+++ client/src/views/Project/Tables/Components/Element/ImageSchemaComponent.vue
@@ -1,6 +1,6 @@
 <script lang="ts">
 import {defineComponent, PropType} from 'vue'
-import {ImageSchema} from "../../types";
+import {ChangeContextEvent, ImageSchema} from "../../types";
 import DataElementSchemaMixin from "../Mixins/DataElementSchemaMixin.ts";
 
 export default defineComponent({
@@ -8,7 +8,8 @@ export default defineComponent({
   mixins: [DataElementSchemaMixin],
   data() {
     return {
-      fileModel: undefined
+      fileModel: undefined,
+      format: undefined as undefined | string
     }
   },
   props: {
@@ -23,9 +24,38 @@ export default defineComponent({
       const reader = new FileReader();
       reader.readAsDataURL(file)
       reader.onload = () => {
-        const bytes = reader?.result?.split(',')[1]
-        this.doWrite({type: 'bytes', value: bytes})
+        const result = reader?.result
+        if (result) {
+          if (this.componentSchema.formatReference) {
+            const format = result.split(';')[0].replace(/^(data:)/,"")
+            const event: ChangeContextEvent = {
+              eventType: 'change-context',
+              contextKey: this.componentSchema.formatReference,
+              value: format
+            }
+            this.doEmitEventSchema(event)
+          }
+          const bytes = result.split(',')[1]
+          this.doWrite({type: 'bytes', value: bytes})
+        }
       }
+    },
+    getFormat() {
+      if (this.format) return this.format
+      return 'image/png'
+    }
+  },
+  mounted() {
+    if (this.componentSchema.formatReference)
+      this.format = this.doRenderContextText(this.componentSchema.formatReference)
+  },
+  watch: {
+    componentContext: {
+      handler() {
+        if (this.componentSchema.formatReference)
+          this.format = this.doRenderContextText(this.componentSchema.formatReference)
+      },
+      deep: true
     }
   }
 })
@@ -39,12 +69,12 @@ export default defineComponent({
   />
   <img
       v-if="(model as any)?.value"
-      :src="'data:image/png;base64,'+ (model as any).value"
+      :src="`data:${getFormat()};base64,`+ (model as any).value"
       :height="componentSchema.size ? CONST_SCHEMA_COMPONENT.image_size[componentSchema.size] : CONST_SCHEMA_COMPONENT.image_size_default"
   />
   <img
       v-else-if="model"
-      :src="'data:image/png;base64,'+ model"
+      :src="`data:${getFormat()};base64,`+ model"
       :height="componentSchema.size ? CONST_SCHEMA_COMPONENT.image_size[componentSchema.size] : CONST_SCHEMA_COMPONENT.image_size_default"
   />
 </template>